Comment regarding input for loadChemicalsList added
Main moved to ASAB level, import shortened in main, comment regarding input of loadChemicalsList added

Initial work on the graph generating functionalities. Visualization of the graph and generation of the tubing.csv file, which is the input for the graph generation.

More edges added in tubing.csv in order to connect y-pieces and syringes and represent valves at syringes.

tubing.csv altered to have values as length and diameter, which can be converted to float for all edges; attributes length, diameter and ends added to the edges in the graph

Desgnation of the density attribute of chemicals changed from density20 to density. main.py adjusted correspondingly. Potential algorithm for finding the shortest path in the graph added as a comment.

New chemList generated in order to get new attribute designation.

Create test folder