Adjustment and rearrangement of Arduino code.
Changes of the Arduino_driver and .ino file
    - the classes Arduino and ArduinoRelay are based on Bojing's code
    - the class ArduinoValve inherits from the valveObj class as it needs to have the basic functions as the other valves provide it
    - commenting and testing are pending
    - in the .ino file the digitalRead calls are added to get the actual state of the relay and thus the actual position of the valve

Arduino Valve driver inital version finished
    - the driver for the Arduino, the relay and the valve are done in a first version
    - the driver also includes the option to simulate the arduino
    - the Arduino class, as well as the ArduinoRelay and ArduinoValve classes behave as the non-simulated ones in the funcionalities offered up to now
    - tests are commented out or fixed to have a clean start for including the Arduino valve into the main commented
    - tests, which were commented out are marked with a todo stating that the test needs to be fixed
    - the change of the .is_open attribute in the emulated balance port was added for the closing functionality to have .is_closed in the correct state afterwards

Preparation for higher test coverage started. All functions to be tested are included in the respective test files. graph, helpers, syringes and solutionHandler still need to be checked.

Commenting and testing of the Arduino_driver
    - tests for the basic functionalities of the Arduino_driver are implemented. The cases for unexpected inputs, e.g. in switch_valve_to_position are still pending.
    - comments to the functions in the Arduino_driver are added
    - The import of the config needs to be fixed. It still seems to be not working as expected.

Various sets of tests implemented
    tests for the following modules and fixes, which appeared reasonable in the modules were done:
        - densiVisco driver and action
        - Arduino driver
        - balance driver and action
        - helpers

    densioVisco renamed to be densiVisco

    CetoniDevice_driver required some changes due to the changed input parameters in helpers.saveToFile().

Tests for syringes added

Work on CetoniDevice_driver tests started, not yet finished.

Further tests added for valves and pumps. Changes in other files due to typeCheck having new inputs now.

tests for CetoniDevice_driver implemented. LoadVariable function in helpers added to work with .py files. Other files adjusted accordingly. Tests for CetoniDevice_driver need to be in the order there as having the test of getValvePositions earlier leads to an access violation.

Implementation of tests for graph started.

Further tests for graph added, next is a test for drawGraph.

Continued implementation of tests for graph.py.

Further tests for graph.py implemented. Only findPumps is pending as it either needs to be moved to the CetoniDevice_driver or implemented more generally.

tests for graph finished. Changes in prepareCetoni required due to changes in target syringes.

Tests for solutionHandler.py added. Changes in files with saveToFile, because it now has folder instead of savePath as a parameter.

Tests for solutionHandler.py finished. Implementation of get_volFracs started.

get_volFracs for molPerMol input done. Tests are also working except for get_solutionDensity.

Alteration of mix started to work with solutions. Not yet finished.

New definition of mixtures and tests for solutionHandler.

Arduino valve included and tested.